AICode/requirements.txt
# 核心工具库
requests>=2.31.0      # 网络请求
python-dotenv>=1.0.0  # 环境变量管理 (.env)
termcolor>=2.4.0      # 终端彩色输出

# AI SDKs
google-genai  # Gemini API
openai>=1.10.0              # OpenAI API (也用于通用的 Token 计算)
anthropic>=0.15.0           # Claude API

# 数据处理
pandas>=2.2.0         # 数据分析 (量化必备)
numpy>=1.26.0         # 数学计算
jsonschema>=4.20.0    # JSON 校验 (Tool Definition)
